requirements.txt
simplejson
six
Django==1.9.5
django-redis==4.3.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
PyJWT==1.4.0
redis==2.10.5
requests==2.9.1
